# Improve string encoding performance